Cleanup the core/zip code.
In 2003, in a possibly-misguided attempt to make compression faster,
the entire contents of the built-in custom compression codec were
converted from compilation targets to header files.  This code, which
is about 90% identical to zlib, had continuous additions throughout
the years, including the code to invoke non-zlib based compression.

With this patch, we have:

- ZDeflate.c/ZInflate.c/ZTrees.c/Bits.c/ZIP.h: Code related to the old
  compression functions.
- Bits.h: headers necessary to actually use the old compression functions.
- RZip.{cxx,h}: Modern headers that anything outside this package should
  use in order to invoke the C-style library functions.
- Compression.h: Enums defining the different compression levels.

Additionally, we now force external callers to utilize the proper enum
for the compression type instead of the type-erased int.
