+++ core/lzma/inc/ZipLZMA.h
@@ -9,6 +9,14 @@
  * For the list of contributors see $ROOTSYS/README/CREDITS.             *
  *************************************************************************/
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 void R__zipLZMA(int cxlevel, int *srcsize, char *src, int *tgtsize, char *tgt, int *irep);
 
 void R__unzipLZMA(int *srcsize, unsigned char *src, int *tgtsize, unsigned char *tgt, int *irep);
+
+#ifdef __cplusplus
+}
+#endif
