+++ core/zip/inc/RZip.h
@@ -8,6 +8,11 @@
  * For the licensing terms see $ROOTSYS/LICENSE.                         *
  * For the list of contributors see $ROOTSYS/README/CREDITS.             *
  *************************************************************************/
+#include "Compression.h"
+
+/**
+ * These are definitions of various free functions for the C-style compression routines in ROOT.
+ */
 
 #ifndef ROOT_RZip
 #define ROOT_RZip
@@ -16,8 +21,12 @@ extern "C" unsigned long R__crc32(unsigned long crc, const unsigned char* buf, u
 
 extern "C" unsigned long R__memcompress(char *tgt, unsigned long tgtsize, char *src, unsigned long srcsize);
 
-extern "C" void R__zipMultipleAlgorithm(int cxlevel, int *srcsize, char *src, int *tgtsize, char *tgt, int *irep, int compressionAlgorithm);
+extern "C" void R__zipMultipleAlgorithm(int cxlevel, int *srcsize, char *src, int *tgtsize, char *tgt, int *irep, ROOT::ECompressionAlgorithm);
 
+/**
+ * This is a historical definition, prior to ROOT supporting multiple algorithms in a single file.  Use
+ * R__zipMultipleAlgorithm instead.
+ */
 extern "C" void R__zip(int cxlevel, int *srcsize, char *src, int *tgtsize, char *tgt, int *irep);
 
 extern "C" void R__unzip(int *srcsize, unsigned char *src, int *tgtsize, unsigned char *tgt, int *irep);
